requirements.txt
pandas==1.5.1
Pillow==9.2.0
